fixed code
