+++ package.json
@@ -18,7 +18,6 @@
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
-    "react-saga": "^0.3.1",
     "react-scripts": "4.0.3",
     "react-thunk": "^1.0.0",
     "redux": "^4.1.2",
