+++ src/components/TodoList/TodoList.scss
@@ -17,15 +17,12 @@
     border-radius: 10px;
     padding: 10px;
     margin-bottom: 5px;
+    background-color: #ffe0db;
+    color: #ff3626;
   }
 
   &__item--checked {
     background-color: #cde4e4;
     color: #459595;
-  }
-
-  &__item--unchecked {
-    background-color: #ffe0db;
-    color: #ff3626;
   }
 }
